+++ src/controllers/controls.ts
@@ -3,7 +3,7 @@ import * as strikeAccess from "../api/strikes";
 import * as meetingAccess from "../api/meetings";
 import * as ogAccess from "../api/ogs";
 import { FCMAdmin } from "../services/fcm";
-import { Liveevent } from "../models/liveevent";
+import { Liveevent } from "../models/liveeventModel";
 
 export const firebaseStatus: RequestHandler = (req, res) => {
   console.log("Firebase status requested");
