+++ src/models/coordsModel.ts
