+++ src/models/ogsModel.ts
