+++ src/models/strikesModel.ts
